requirements.txt
python-dateutil
certifi
cryptography
jinja2
requests
